Intégration des notes et des alternatives d'achat
