+++ src/Vl1.js
@@ -2,6 +2,8 @@ import correspondance from './correspondance-icônes.yaml'
 import vélo from '../vélos/1.yaml'
 import {Card} from './ui'
 import {domain} from './utils'
+import {Markdown} from './utils'
+import {useState} from 'react'
 
 
 export default ({}) => (
@@ -28,25 +30,58 @@ export default ({}) => (
 			<p>{vélo.description}</p>
 			<ul>
 				{Object.entries(vélo.composants).filter(([_, data]) => data != null).map(
-					([composant, {modèle, marque, prix, url}]) => (
-						<li>
-							<div css="text-align: center; max-width: 6rem; margin-right: 1rem">
-								<img
-									css="width: 5rem" src={'composants/' + correspondance[composant] + '.svg'}
-								/>
-								<div css="text-transform: uppercase; font-size: 85%; color: #666">
-									{composant}
-								</div>
-							</div>
-							<Card><div>
-								<span css="font-size: 90%; font-weight: bold; margin-right: .4rem">{marque}</span>
-								<span>{modèle}</span></div>
-								<div>{prix} {domain(url) && <span>sur <a href={url}>{domain(url)}</a></span>}</div>
-							</Card>
-						</li>
-					)
+					item => <Composant item={item} />
 				)}
 			</ul>
 		</div >
 	</div >
 )
+
+const Composant = ({item:
+	[composant, data]
+}) => {
+	const note = data.note,
+		chosen = data.alternatives ? data.alternatives[0] : data
+		, {modèle, marque} = chosen
+		, sold = chosen.achat ? chosen.achat[0] : chosen,
+		{prix, url} = sold
+
+
+	return (<li>
+		<div css="text-align: center; max-width: 6rem; margin-right: 1rem">
+			<img
+				css="width: 5rem" src={'composants/' + correspondance[composant] + '.svg'}
+			/>
+			<div css="text-transform: uppercase; font-size: 85%; color: #666">
+				{composant}
+			</div>
+		</div>
+		<div>
+			<Note data={note} />
+			<Card><div>
+				<span css="font-size: 90%; font-weight: bold; margin-right: .4rem">{marque}</span>
+				<span>{modèle}</span></div>
+				<div>{prix} {domain(url) && <span>sur <a href={url}>{domain(url)}</a></span>}</div>
+			</Card>
+		</div>
+	</li>
+	)
+}
+
+const Note = ({data}) => {
+
+	const [open, setOpen] = useState(false)
+	if (!data) return null
+	const [intro, ...rest] = data.split('\n')
+
+	return (
+		<div>
+			{open}
+			<Markdown source={open ? data : intro} />
+			<p>
+				<button onClick={() => console.log('yaya') || setOpen(!open)}> {open ? 'Réduire' : 'Lire plus'}</button>
+			</p>
+		</div>
+
+	)
+}
