+++ vélos/1.yaml
@@ -36,7 +36,6 @@ composants:
     matériaux: 
       - aluminium
   étriers de frein: 
-    inclus: disque
     note: |
       Les freins à disque sont incontournables pour le gravel, pour plusieurs raisons pratiques: 
       - il est tout simplement très difficile de trouver un cadre gravel (donc qui au moins accueille des pneus de diamètre > 30mm) à freins sur jante
@@ -48,43 +47,44 @@ composants:
 
       Pour que le prix et la difficulté de montage restent contenus, nous partons sur des freins disque mécaniques.
 
-      marque: Avid
-      modèle: BB7
-      note: |
-        Les commentaires sur ces freins sont plutôt élogieux. Attention, avant/arrière semblent différents.
+    alternatives:
+      - 
+        marque: Avid
+        modèle: BB7
+        note: |
+          Les commentaires sur ces freins sont plutôt élogieux. Attention, avant/arrière semblent différents.
 
-        > Properly adjusted BB7's are awesome brakes. Key = Properly adjusted. The disc shouldn't be centered between the pads, but instead a slightly larger gap on the inside. This will give you the great "feel" that BB7s are known for.
-        Semblent être en fixation flat mount ✅.
+          > Properly adjusted BB7's are awesome brakes. Key = Properly adjusted. The disc shouldn't be centered between the pads, but instead a slightly larger gap on the inside. This will give you the great "feel" that BB7s are known for.
+          Semblent être en fixation flat mount ✅.
 
-      achat: 
-        - url: https://www.materiel-velo.com/1858732-etrier-velo-route-avid-bb7-disque-mecanique.html
-          prix: 54€
-        - url: https://www.probikeshop.fr/frein-arriere-a-disque-sram-bb7-road/117408.html
-          prix: 85
-        - url: https://www.materiel-velo.com/1858732-etrier-velo-route-avid-bb7-disque-mecanique.html
-          prix: 54
-      inclus: 
-        - disques
-      alternatives:
-        - url: https://www.decathlon.fr/p/kit-frein-a-disque-mecanique-arriere/_/R-p-202955
-        - url: https://www.decathlon.fr/p/kit-frein-mecanique-hayes-mx-5-disque-160mm/_/R-p-196567
-        - marque: Décathlon
-          modèle: KIT FREIN A DISQUE MECANIQUE
-          url: https://www.decathlon.fr/p/kit-frein-a-disque-mecanique/_/R-p-141530
-          prix: 29.90€
-          note: Alternative bas coût. Quel format, postmount, flatmount ? En tout cas vendu avec adaptateur
-        - marque: TRP 
-          modèle: Spyre
-          rupture: partout
-          prix: 70.58€
-          url: https://www.bike24.com/p2225128.html?menu=1000,4,322
-          note: |
-            J'ai pu essayer ces freins, ils ne sont pas incroyables mais font bien l'affaire pour un prix contenu.
-            
-            A noter : les plaquettes bougent des deux côtés du disque, ce qui semble mieux pour éviter de le voiler. 
-
-            D'après [ces retours](https://www.reddit.com/r/gravelcycling/comments/gutw00/mech_disc_brake_suggestions_avid_bb5bb7_trp_spyre/), ils sont meilleurs que les BB7.
+        achat: 
+          - url: https://www.materiel-velo.com/1858732-etrier-velo-route-avid-bb7-disque-mecanique.html
+            prix: 54€
+          - url: https://www.probikeshop.fr/frein-arriere-a-disque-sram-bb7-road/117408.html
+            prix: 85
+          - url: https://www.materiel-velo.com/1858732-etrier-velo-route-avid-bb7-disque-mecanique.html
+            prix: 54
+        inclus: 
+          - disques
+      - url: https://www.decathlon.fr/p/kit-frein-a-disque-mecanique-arriere/_/R-p-202955
+      - url: https://www.decathlon.fr/p/kit-frein-mecanique-hayes-mx-5-disque-160mm/_/R-p-196567
+      - marque: Décathlon
+        modèle: KIT FREIN A DISQUE MECANIQUE
+        url: https://www.decathlon.fr/p/kit-frein-a-disque-mecanique/_/R-p-141530
+        prix: 29.90€
+        note: Alternative bas coût. Quel format, postmount, flatmount ? En tout cas vendu avec adaptateur
+      - marque: TRP 
+        modèle: Spyre
+        rupture: partout
+        prix: 70.58€
+        url: https://www.bike24.com/p2225128.html?menu=1000,4,322
+        note: |
+          J'ai pu essayer ces freins, ils ne sont pas incroyables mais font bien l'affaire pour un prix contenu.
           
+          A noter : les plaquettes bougent des deux côtés du disque, ce qui semble mieux pour éviter de le voiler. 
+
+          D'après [ces retours](https://www.reddit.com/r/gravelcycling/comments/gutw00/mech_disc_brake_suggestions_avid_bb5bb7_trp_spyre/), ils sont meilleurs que les BB7.
+        
   cables et gaines: 
     note: |
       Prendre de bons trucs, genre Jagwire kevlar, pour avoir une bonne qualité de freinage.
